chore(tree): Move tree related files to root of the lib
