+++ crates/blockchain-tree/src/block_indices.rs
