+++ crates/blockchain-tree/src/externals.rs
