+++ crates/blockchain-tree/src/lib.rs
@@ -4,17 +4,28 @@
     no_crate_inject,
     attr(deny(warnings, rust_2018_idioms), allow(dead_code, unused_variables))
 ))]
-
-//! Reth executor executes transaction in block of data.
-
-pub mod eth_dao_fork;
+//! Implementation of the [BlockchainTree]
 
 /// Execution result types.
 pub use reth_provider::post_state;
 
 pub mod blockchain_tree;
-pub use blockchain_tree::*;
+pub use blockchain_tree::{BlockHashes, BlockchainTree};
 
-#[cfg(any(test, feature = "test-utils"))]
-/// Common test helpers for mocking out executor and executor factory
-pub mod test_utils;
+pub mod block_indices;
+pub use block_indices::BlockIndices;
+
+pub mod chain;
+pub use chain::AppendableChain;
+
+pub mod config;
+pub use config::BlockchainTreeConfig;
+
+pub mod externals;
+pub use externals::TreeExternals;
+
+pub mod shareable;
+pub use shareable::ShareableBlockchainTree;
+
+pub mod post_state_data;
+pub use post_state_data::{PostStateData, PostStateDataRef};
