--- crates/blockchain-tree/src/test_utils/mod.rs
@@ -1,5 +1,0 @@
-mod executor;
-pub use executor::*;
-
-mod factory;
-pub use factory::*;
