+++ crates/revm/src/lib.rs
@@ -24,3 +24,6 @@ pub use reth_revm_primitives::*;
 
 /// Re-export everything
 pub use revm;
+
+/// Etereum DAO hardfork state change data.
+pub mod eth_dao_fork;
