+++ crates/storage/provider/src/test_utils/mod.rs
@@ -1,8 +1,10 @@
 pub mod blocks;
 mod events;
+mod executor;
 mod mock;
 mod noop;
 
 pub use events::TestCanonStateSubscriptions;
+pub use executor::{TestExecutor, TestExecutorFactory};
 pub use mock::{ExtendedAccount, MockEthProvider};
 pub use noop::NoopProvider;
